[!!!][FEATURE] Preload file records in Storage
Currently, the drivers load file data from the physical storage for each
and every file. For large numbers of files, this might be quite
inefficient, especially for getting things like the mimetype. Most of
this information is also already present in the database.

This patch changes the default behavior of the Storage to load the
records of all files when getFileList() is called and put the