[!!!][FEATURE] Hide special files in file and folder lists
Currently, even dot-files (.htaccess etc.) are shown in file lists. It
is not possible to exclude them, except some weird UI-wise custom
filtering.

To enable proper filtering, this commit adds a default filter hidden
files to the getFileList()/getFolderList() methods of the drivers, plus
a method in the storage to

Resolves: #33648